Developing profiles module

+++ users/static/users/js/app/listing.js
@@ -1,24 +1,27 @@
 function deleteUser(user_id){
-    $.ajax({
-        url: user_delete_url + user_id,
-        method: 'DELETE',
-        async: false,
-        dataType: 'json',
-        data: {},
-        beforeSend: function(){},
-        success: function(result){
-            $('#modalMessage').html("El usuario ha sido eliminado con exito");
-            $('#successModal').modal('toggle');
-            $('#successModal').modal({backdrop:'static', keyboard:false}); 
-            setTimeout(function(){ 
-                location.reload();
-            }, 2000);
-        },
-        error: function (result, request, status, error){},
-        complete: function(){
-            singleOperationRestriction=false
-        },
-    });
+    if(!singleOperationRestriction){
+        singleOperationRestriction = true;
+        $.ajax({
+            url: delete_url + user_id,
+            method: 'DELETE',
+            async: false,
+            dataType: 'json',
+            data: {},
+            beforeSend: function(){},
+            success: function(result){
+                $('#modalMessage').html("El usuario ha sido eliminado con exito");
+                $('#successModal').modal('toggle');
+                $('#successModal').modal({backdrop:'static', keyboard:false}); 
+                setTimeout(function(){ 
+                    location.reload();
+                }, 2000);
+            },
+            error: function (result, request, status, error){},
+            complete: function(){
+                singleOperationRestriction=false
+            },
+        });
+    }
 }
 
 function deleteConfirmation(user_id){
@@ -27,6 +30,38 @@ function deleteConfirmation(user_id){
     });
     $('#idi').html(user_id);
     $('#cautionModal').modal('toggle');
+}
+
+function toogleUserState(user_id, square_element){
+    if(!singleOperationRestriction){
+        singleOperationRestriction = true;
+        $.ajax({
+            url: toggle_url + user_id,
+            method: 'POST',
+            async: false,
+            dataType: 'json',
+            data: {},
+            beforeSend: function(){},
+            success: function(result){
+                if(result.success){
+                    if(square_element.hasClass('fa-square')){
+                        square_element.removeClass('fa-square');
+                        square_element.addClass('fa-check-square');
+                        SoftNotification.show("Usuario activado con exito");
+                    }
+                    else if(square_element.hasClass('fa-check-square')){
+                        square_element.removeClass('fa-check-square');
+                        square_element.addClass('fa-square');
+                        SoftNotification.show("Usuario desactivado con exito");
+                    }
+                }
+            },
+            error: function (result, request, status, error){},
+            complete: function(){
+                singleOperationRestriction=false
+            },
+        });
+    }
 }
 
 $(document).ready(function() {
@@ -72,7 +107,7 @@ $(document).ready(function() {
             $('.row_object_id').dblclick(function(event){
                 var target = $(event.delegateTarget);
                 var id = target.attr('user_id');
-                window.location = user_update_url + id;
+                window.location = update_url + id;
             });
 
             $('.fa-trash, .fa-square, .fa-check-square').click(function(event){
@@ -84,21 +119,20 @@ $(document).ready(function() {
                 if(target.hasClass('fa-trash')){
                     deleteConfirmation(user_id);
                 }
-                else if(target.hasClass('fa-square')){
-                    console.log('square');
-                }
-                else if(target.hasClass('fa-check-square')){
-                    console.log('check');
+                else if(target.hasClass('fa-square') || target.hasClass('fa-check-square')){
+                    toogleUserState(user_id, target);
                 }
             });
 
             $('.row_object_id').css('cursor', 'pointer'); 
+
+            $(function () {
+                $('[data-toggle="popover"]').popover()
+            });
         }
     });
 
     $('#listing tbody').on( 'click', 'tr', function () {
         $(this).toggleClass('selected');
     });
-
-
 } );
